requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.5.0
asyncio==3.4.3
asyncmy==0.2.9
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Faker==29.0.0
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
